[sbt] remove legacy test SbtVersionDetectorTest.testMockLauncherWithoutSbtBootProperties
Version in the manifest file does not represent the version of the SBT.
Sbt version can only be in the boot properties of the launcher jar.
But it's a fallback case anyway.
Ideally the version should be specified in the `project/build.properties` file or in the new project wizard