[internal build] build attributed JPS classpath and attach sources instead of simply using unmanagedJars #SCL-17857 fixed
Previously, the jar files added with `unmanagedJars` were added as a generic IntelliJ library named "sbt: unmanaged-jars." The issue was that the name is not nice and that it didn't have any sources attached. This setup broke source navigation in the K2 mode (SCL-17857). This is because Kotlin plugin used the first found jar file that contains kotlin classes (it found some lib/app*.jar from the IntelliJ SDK root that is included in the JPS classpath).

The updated approach registers JARs similarly to IntelliJ SDK libraries. It assigns them distinct names and attaches IntelliJ SDK sources.