[psi] do not reindex file if it has a wrong stub type, just log the error
Method FileBasedIndex.requestReindex should not be called in getStub, it's a wrong lifecycle moment to do so.
In some edge cases it might lead to infinite cyclic reindexing.
So far I only noticed it during the local development with some state of code.
But this is too risky to leave it here.